refactor: rename containers folder
