+++ public/sitemap.xml
@@ -1,15 +1,15 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:news="http://www.google.com/schemas/sitemap-news/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml" xmlns:mobile="http://www.google.com/schemas/sitemap-mobile/1.0" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1" xmlns:video="http://www.google.com/schemas/sitemap-video/1.1">
-<url><loc>https://morethan-log.vercel.app</loc><lastmod>2023-07-09T13:47:57.837Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://morethan-log.vercel.app/project-feedback</loc><lastmod>2023-07-09T13:47:57.837Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://morethan-log.vercel.app/2022-end-2023-start</loc><lastmod>2023-07-09T13:47:57.837Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://morethan-log.vercel.app/how-to-generate-dynamic-og-image</loc><lastmod>2023-07-09T13:47:57.837Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://morethan-log.vercel.app/zext-project-review</loc><lastmod>2023-07-09T13:47:57.837Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://morethan-log.vercel.app/hot-to-make-sitemap-in-next-js</loc><lastmod>2023-07-09T13:47:57.837Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://morethan-log.vercel.app/apollo-gettings-started-mutate</loc><lastmod>2023-07-09T13:47:57.837Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://morethan-log.vercel.app/apollo-gettings-started-query</loc><lastmod>2023-07-09T13:47:57.837Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://morethan-log.vercel.app/about-graphql</loc><lastmod>2023-07-09T13:47:57.837Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://morethan-log.vercel.app/how-to-apply-ga-to-next.js</loc><lastmod>2023-07-09T13:47:57.837Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://morethan-log.vercel.app/about-opensource-license</loc><lastmod>2023-07-09T13:47:57.837Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://morethan-log.vercel.app/about</loc><lastmod>2023-07-09T13:47:57.837Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://morethan-log.vercel.app</loc><lastmod>2023-07-13T13:50:33.547Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://morethan-log.vercel.app/project-feedback</loc><lastmod>2023-07-13T13:50:33.547Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://morethan-log.vercel.app/2022-end-2023-start</loc><lastmod>2023-07-13T13:50:33.547Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://morethan-log.vercel.app/how-to-generate-dynamic-og-image</loc><lastmod>2023-07-13T13:50:33.547Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://morethan-log.vercel.app/zext-project-review</loc><lastmod>2023-07-13T13:50:33.547Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://morethan-log.vercel.app/hot-to-make-sitemap-in-next-js</loc><lastmod>2023-07-13T13:50:33.547Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://morethan-log.vercel.app/apollo-gettings-started-mutate</loc><lastmod>2023-07-13T13:50:33.547Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://morethan-log.vercel.app/apollo-gettings-started-query</loc><lastmod>2023-07-13T13:50:33.547Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://morethan-log.vercel.app/about-graphql</loc><lastmod>2023-07-13T13:50:33.547Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://morethan-log.vercel.app/how-to-apply-ga-to-next.js</loc><lastmod>2023-07-13T13:50:33.547Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://morethan-log.vercel.app/about-opensource-license</loc><lastmod>2023-07-13T13:50:33.547Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://morethan-log.vercel.app/about</loc><lastmod>2023-07-13T13:50:33.547Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
 </urlset>
