+++ src/routes/Detail/components/PageDetail/index.tsx
