+++ src/routes/Detail/components/PostDetail/PostFooter.tsx
