+++ src/routes/Detail/components/PostDetail/index.tsx
