+++ src/routes/Detail/hooks/useMermaidEffect.ts
