+++ src/routes/Detail/index.tsx
